requirements.txt
opencv-python
ultralytics
requests
face_recognition
facenet-pytorch
torch
Pillow
